requirements.txt
requests
pandas
numpy
confluent-kafka
pydantic
websocket
websocket-client
aiokafka
cctx
kafka-python-ng
mmh3